seamai/requirements.txt
numpy==1.24.4
opencv-python==4.8.0.74
#torch==2.0.1
#torchvision==0.15.2
torchview==0.2.6
#torchmetrics==0.11.4
#pytorchvideo==0.1.5
#torchinfo==1.8.0
tqdm==4.66.1
pandas==2.0.3
matplotlib==3.7.3
pathlib==1.0.1
imutils==0.5.4
scikit-learn==1.3.0
scikit-image==0.21.0
wandb==0.15.12
psutil==5.9.5
#gpiozero==2.0
GPUtil==1.4.0
py-cpuinfo==9.0.0
moviepy==1.0.3
einops==0.7.0
